+++ src/components/Table/components/AggregatedCellRenderer/AggregatedCellRenderer.tsx
@@ -1,30 +1,31 @@
-import { FormattedValueDisplay } from '@grafana/ui';
+import { FormattedValueDisplay, useTheme2 } from '@grafana/ui';
 import { CellContext } from '@tanstack/react-table';
 import React from 'react';
 
-import { TEST_IDS } from '@/constants';
-import { CellAggregation } from '@/types';
+import { AGGREGATION_TYPES_WITH_DISPLAY_PROCESSOR, TEST_IDS } from '@/constants';
+import { CellType } from '@/types';
 
 /**
  * Properties
  */
-interface Props extends CellContext<unknown, unknown> {}
-
-/**
- * Aggregation Type acceptable for display processor
- */
-export const acceptableAggregationType = [
-  CellAggregation.MIN,
-  CellAggregation.MAX,
-  CellAggregation.MEAN,
-  CellAggregation.MEDIAN,
-  CellAggregation.SUM,
-];
+interface Props extends CellContext<unknown, unknown> {
+  /**
+   * Bg Color
+   *
+   * @type {string}
+   */
+  bgColor?: string;
+}
 
 /**
  * Aggregated Cell Renderer
  */
-export const AggregatedCellRenderer: React.FC<Props> = ({ renderValue, column }) => {
+export const AggregatedCellRenderer: React.FC<Props> = ({ renderValue, column, bgColor }) => {
+  /**
+   * Theme
+   */
+  const theme = useTheme2();
+
   /**
    * No meta
    */
@@ -39,11 +40,23 @@ export const AggregatedCellRenderer: React.FC<Props> = ({ renderValue, column })
   /**
    * Use Display Processor
    */
-  if (field.display && acceptableAggregationType.includes(config.aggregation)) {
+  if (field.display && AGGREGATION_TYPES_WITH_DISPLAY_PROCESSOR.includes(config.aggregation)) {
     const displayValue = field.display(value);
 
+    let color = 'inherit';
+
+    if (displayValue.color) {
+      color = displayValue.color;
+    }
+
     return (
-      <span {...TEST_IDS.aggregatedCellRenderer.root.apply()}>
+      <span
+        style={{
+          color:
+            config.type === CellType.COLORED_TEXT ? color : bgColor ? theme.colors.getContrastText(bgColor) : 'inherit',
+        }}
+        {...TEST_IDS.aggregatedCellRenderer.root.apply()}
+      >
         <FormattedValueDisplay value={displayValue} />
       </span>
     );